Encode http content as UTF-8 before regex matching
This change fixes issue https://github.com/Valarissa/transgress/issues/3
where visiting www.google.com causes a 500 error instead of showing the
"IRRELEVANT" page.

The issue appears to be caused by treating non UTF-8 content as UTF-8. A
question related to this was asked on Stack Overflow [1] and the answer
here [2] suggests encoding as UTF-8 with `encode!` so that is what we do
here.

[1] http://stackoverflow.com/q/2982677/119768
[2] http://stackoverflow.com/a/8873922/119768